refactor(schema): move `search` schema file to `@repo/schema`
