+++ packages/schema/src/search.ts
